Update to TicTacToe app
I've tried to do the refactoring you suggested, but I ended up with a
different result then the proposed one. I hope the code is not as ugly
as it was.

Unfortunately, it does not work, as I get a NullPointer error at line 58
in the main activity. After debugging I can confirm that the lines 46-54
have no effect whatsoever, the buttonMatrix contains null values after
the assignments so when the app tries to call
buttonMatrix[i][j].resetContent(); it obviously crashes.

The problem is I have no idea why the assignments don't work.

And as a second thought the onClickListener still seems ugly, but I had
no better idea.