Project database implementation
I think there will be no need for separate Students and Teachers tables,
because al the data from these tables from the 1st version are now
present in other tables, so only the Persons table remains which will be
connected to all teh other tables through the personID column. I hope it
is OK this way.