Tetris
New pieces are generated and the pieces get placed on the board.
I have some trouble with rotating the pieces.
I am done with the logic of rotating the pieces (rotatePiece method in
class Shape). I am not sure how to make the connection between this
method and the UI.
Also, I have made some of the changes you wanted me to make.