Tetris final changes
Thank you!
I have made the final changes:
deleted everything extra and arranged the code. Also, I changed the
color of the board and removed the unnecessary borders.
Concerning that method, the idea was to also check for other pieces
already on the board, that is why I thought checking for the color would
be the easiest solution to that problem.